🧱 🐛 perform display preparation processing within the geometry class
